requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiohttp-retry==2.9.1
aiosignal==1.4.0
amqp==5.3.1
annotated-types==0.7.0
anyio==4.10.0
asgiref==3.9.1
asttokens==3.0.0
attrs==25.3.0
audioop-lts==0.2.2
autobahn==24.4.2
Automat==25.4.16
beautifulsoup4==4.13.5
billiard==4.2.2
binary==1.0.2
black==25.9.0
blis==1.3.0
boto3==1.40.35
botocore==1.40.35
catalogue==2.0.10
celery==5.5.3
certifi==2025.8.3
cffi==2.0.0
cfgv==3.4.0
channels==4.3.1
channels_redis==4.3.0
charset-normalizer==3.4.3
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cloudpathlib==0.22.0
colorama==0.4.6
confection==0.1.5
constantly==23.10.4
coverage==7.10.6
cron_descriptor==2.0.6
cryptography==46.0.1
cymem==2.0.11
daphne==4.2.1
decorator==5.2.1
distlib==0.4.0
distro==1.9.0
dj-database-url==3.0.1
Django==5.2.6
django-allauth==65.11.2
django-celery-beat==2.8.1
django-cors-headers==4.9.0
django-debug-toolbar==6.0.0
django-elasticsearch-dsl==8.2
django-environ==0.12.0
django-extensions==4.1
django-filter==25.1
django-location-field==2.7.3
django-oauth-toolkit==3.0.1
django-phonenumber-field==8.1.0
django-ratelimit==4.1.0
django-redis==6.0.0
django-seed==0.3.1
django-storages==1.14.6
django-timezone-field==7.1
django_celery_results==2.6.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
docopt==0.6.2
drf-spectacular==0.28.0
drf-spectacular-sidecar==2025.9.1
elastic-transport==8.17.1
elasticsearch==8.19.1
elasticsearch-dsl==8.15.4
et_xmlfile==2.0.0
executing==2.2.1
factory_boy==3.3.3
Faker==37.8.0
filelock==3.13.1
flake8==7.3.0
flower==2.0.1
frozenlist==1.7.0
fsspec==2024.6.1
geographiclib==2.1
geoip2==5.1.0
geopy==2.4.1
gTTS==2.5.4
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.35.0
humanize==4.13.0
hyperlink==21.0.0
identify==2.6.14
idna==3.10
incremental==24.7.2
inflection==0.5.1
iniconfig==2.1.0
ipython==9.5.0
ipython_pygments_lexers==1.1.1
isort==6.0.1
jedi==0.19.2
Jinja2==3.1.4
jiter==0.11.0
jmespath==1.0.1
joblib==1.5.2
Js2Py==0.74
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jwcrypto==1.5.6
kombu==5.5.4
langcodes==3.5.0
langdetect==1.0.9
language_data==1.3.0
lxml==6.0.1
marisa-trie==1.3.1
markdown-it-py==4.0.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
maxminddb==2.8.2
mccabe==0.7.0
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.1.1
multidict==6.6.4
murmurhash==1.0.13
mypy_extensions==1.1.0
networkx==3.3
nltk==3.9.1
nodeenv==1.9.1
numpy==2.3.3
oauthlib==3.3.1
openai==1.108.1
opencv-python==4.11.0.86
openpyxl==3.1.5
packaging==25.0
pandas==2.3.2
parso==0.8.5
pathspec==0.12.1
phonenumbers==9.0.14
pillow==11.3.0
pipwin==0.5.2
platformdirs==4.4.0
pluggy==1.6.0
pre_commit==4.3.0
preshed==3.0.10
prometheus_client==0.23.1
prompt_toolkit==3.0.52
propcache==0.3.2
psycopg==3.2.10
psycopg-binary==3.2.10
psycopg2-binary==2.9.10
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycodestyle==2.14.0
pycparser==2.23
pydantic==2.11.9
pydantic_core==2.33.2
pydub==0.25.1
pyflakes==3.4.0
Pygments==2.19.2
pyjsparser==2.7.1
PyJWT==2.10.1
pyOpenSSL==25.3.0
PyPDF2==3.0.1
PyPrind==2.11.3
pySmartDL==1.3.4
pytest==8.4.2
pytest-django==4.11.1
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-docx==1.2.0
python-dotenv==1.1.1
python-http-client==3.3.7
python-magic==0.4.27
pytokens==0.1.10
pytz==2025.2
PyYAML==6.0.2
redis==6.4.0
referencing==0.36.2
regex==2025.9.18
requests==2.32.5
rich==14.1.0
rpds-py==0.27.1
s3transfer==0.14.0
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.2
sendgrid==6.12.5
sentence-transformers==5.1.0
sentry-sdk==2.38.0
service-identity==24.2.0
setuptools==80.9.0
shapely==2.1.1
shellingham==1.5.4
six==1.17.0
smart_open==7.3.1
sniffio==1.3.1
soupsieve==2.8
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SpeechRecognition==3.14.3
sqlparse==0.5.3
srsly==2.5.1
stack-data==0.6.3
standard-aifc==3.13.0
standard-chunk==3.13.0
sympy==1.13.3
textblob==0.19.0
thinc==8.3.6
threadpoolctl==3.6.0
tokenizers==0.22.1
toposort==1.10
torch==2.8.0
torchaudio==2.8.0
torchvision==0.23.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.56.2
twilio==9.8.1
Twisted==25.5.0
txaio==25.6.1
typer==0.19.1
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.5.0
uuid==1.30
vine==5.1.0
virtualenv==20.34.0
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
Werkzeug==3.1.3
whitenoise==6.11.0
wrapt==1.17.3
yarl==1.20.1
zope.interface==8.0
